Remove from the OMTF constructor all the variables and use all the kargs. Move also all auxiliary functions to the converter.py file
